Moved components used by explorer to common directory
