+++ src/pages/details/azureDetails/azureDetails.tsx
@@ -6,11 +6,12 @@ import { tagPrefix } from 'api/queries/query';
 import { AzureReport } from 'api/reports/azureReports';
 import { ReportPathsType, ReportType } from 'api/reports/report';
 import { AxiosError } from 'axios';
-import { addQueryFilter, getGroupByTagKey, removeQueryFilter } from 'pages/details/common/detailsUtils';
-import { ExportModal } from 'pages/details/components/export/exportModal';
+import { ExportModal } from 'pages/components/export/exportModal';
 import Loading from 'pages/state/loading';
 import NoProviders from 'pages/state/noProviders';
 import NotAvailable from 'pages/state/notAvailable';
+import { getGroupByTagKey } from 'pages/utils/groupBy';
+import { addQueryFilter, removeQueryFilter } from 'pages/utils/query';
 import React from 'react';
 import { WithTranslation, withTranslation } from 'react-i18next';
 import { connect } from 'react-redux';
