+++ src/pages/details/ocpBreakdown/costOverview.tsx
@@ -1,8 +1,8 @@
 import { CostOverviewBase } from 'pages/details/components/costOverview/costOverviewBase';
 import { WithTranslation, withTranslation } from 'react-i18next';
 import { connect } from 'react-redux';
+import { ocpCostOverviewSelectors } from 'store/breakdown/costOverview/ocpCostOverview';
 import { createMapStateToProps } from 'store/common';
-import { ocpCostOverviewSelectors } from 'store/costOverview/ocpCostOverview';
 
 interface CostOverviewStateProps {
   widgets: number[];
