+++ src/pages/explorer/explorer.tsx
@@ -7,12 +7,13 @@ import { getUserAccessQuery } from 'api/queries/userAccessQuery';
 import { Report } from 'api/reports/report';
 import { UserAccess, UserAccessType } from 'api/userAccess';
 import { AxiosError } from 'axios';
-import { addQueryFilter, getGroupByOrg, getGroupByTagKey, removeQueryFilter } from 'pages/details/common/detailsUtils';
-import { ExportModal } from 'pages/details/components/export/exportModal';
+import { ExportModal } from 'pages/components/export/exportModal';
 import Loading from 'pages/state/loading';
 import NoData from 'pages/state/noData';
 import NoProviders from 'pages/state/noProviders';
 import NotAvailable from 'pages/state/notAvailable';
+import { getGroupByOrg, getGroupByTagKey } from 'pages/utils/groupBy';
+import { addQueryFilter, removeQueryFilter } from 'pages/utils/query';
 import React from 'react';
 import { WithTranslation, withTranslation } from 'react-i18next';
 import { connect } from 'react-redux';
