+++ src/pages/explorer/explorerHeader.tsx
@@ -5,8 +5,8 @@ import { getQuery, parseQuery, Query } from 'api/queries/query';
 import { getUserAccessQuery } from 'api/queries/userAccessQuery';
 import { UserAccess, UserAccessType } from 'api/userAccess';
 import { AxiosError } from 'axios';
-import { GroupBy } from 'pages/details/components/groupBy/groupBy';
-import { Perspective } from 'pages/overview/perspective';
+import { GroupBy } from 'pages/components/groupBy/groupBy';
+import { Perspective } from 'pages/components/perspective/perspective';
 import React from 'react';
 import { WithTranslation, withTranslation } from 'react-i18next';
 import { connect } from 'react-redux';
