+++ src/pages/utils/groupBy.ts
@@ -1,0 +1,36 @@
+import { orgUnitIdKey, Query, tagPrefix } from 'api/queries/query';
+
+export const getGroupById = (query: Query) => {
+  const groupBys = query && query.group_by ? Object.keys(query.group_by) : [];
+  return groupBys.find(key => key !== orgUnitIdKey);
+};
+
+export const getGroupByValue = (query: Query) => {
+  const groupById = getGroupById(query);
+  return groupById ? query.group_by[groupById] : undefined;
+};
+
+export const getGroupByOrg = (query: Query) => {
+  let groupByOrg;
+
+  for (const groupBy of Object.keys(query.group_by)) {
+    if (groupBy === orgUnitIdKey) {
+      groupByOrg = query.group_by[orgUnitIdKey];
+      break;
+    }
+  }
+  return groupByOrg;
+};
+
+export const getGroupByTagKey = (query: Query) => {
+  let groupByTagKey;
+
+  for (const groupBy of Object.keys(query.group_by)) {
+    const tagIndex = groupBy.indexOf(tagPrefix);
+    if (tagIndex !== -1) {
+      groupByTagKey = groupBy.substring(tagIndex + tagPrefix.length) as any;
+      break;
+    }
+  }
+  return groupByTagKey;
+};
