+++ src/pages/utils/query.ts
@@ -1,29 +1,4 @@
-import { orgUnitIdKey, Query, tagPrefix } from 'api/queries/query';
-
-export const getGroupByOrg = (query: Query) => {
-  let groupByOrg;
-
-  for (const groupBy of Object.keys(query.group_by)) {
-    if (groupBy === orgUnitIdKey) {
-      groupByOrg = query.group_by[orgUnitIdKey];
-      break;
-    }
-  }
-  return groupByOrg;
-};
-
-export const getGroupByTagKey = (query: Query) => {
-  let groupByTagKey;
-
-  for (const groupBy of Object.keys(query.group_by)) {
-    const tagIndex = groupBy.indexOf(tagPrefix);
-    if (tagIndex !== -1) {
-      groupByTagKey = groupBy.substring(tagIndex + tagPrefix.length) as any;
-      break;
-    }
-  }
-  return groupByTagKey;
-};
+import { Query } from 'api/queries/query';
 
 export const addQueryFilter = (query: Query, filterType: string, filterValue: string) => {
   const newQuery = { ...JSON.parse(JSON.stringify(query)) };
